+++ mirrord/layer/src/lib.rs
@@ -10,8 +10,12 @@
 #![feature(try_trait_v2)]
 #![feature(try_trait_v2_residual)]
 #![feature(trait_alias)]
+#![feature(c_size_t)]
+#![feature(pointer_byte_offsets)]
 
 extern crate alloc;
+extern crate core;
+
 use std::{
     collections::VecDeque,
     net::{IpAddr, Ipv4Addr, Ipv6Addr, SocketAddr},