+++ mirrord/layer/tests/common/mod.rs
@@ -339,6 +339,7 @@ pub enum Application {
     EnvBashCat,
     NodeFileOps,
     GoDir,
+    GoDirBypass,
     Go19Issue834,
     Go18Issue834,
 }
@@ -385,6 +386,7 @@ impl Application {
             Application::GoDir => String::from("tests/apps/dir_go/dir_go"),
             Application::Go19Issue834 => String::from("tests/apps/issue834/19"),
             Application::Go18Issue834 => String::from("tests/apps/issue834/18"),
+            Application::GoDirBypass => String::from("tests/apps/dir_go_bypass/dir_go_bypass"),
         }
     }
 
@@ -427,6 +429,7 @@ impl Application {
             Application::Go18Issue834 => vec![],
             Application::RustFileOps => vec![],
             Application::EnvBashCat => vec![],
+            Application::GoDirBypass => vec![],
         }
     }
 
@@ -443,6 +446,7 @@ impl Application {
             | Application::NodeFileOps
             | Application::Go19Issue834
             | Application::Go18Issue834
+            | Application::GoDirBypass
             | Application::GoDir => {
                 unimplemented!("shouldn't get here")
             }