+++ CHANGELOG.md
@@ -7,9 +7,15 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
+### Added
+
+- Support for Go's `os.ReadDir` on Linux (by hooking the `getdents64` syscall). Part of
+  [#120](https://github.com/metalbear-co/mirrord/issues/120).
+
 ### Fixed
 
-- mirrord-agent: Handle HTTP upgrade requests when the stealer feature is enabled (with HTTP traffic) PR [#973](https://github.com/metalbear-co/mirrord/pull/973).
+- mirrord-agent: Handle HTTP upgrade requests when the stealer feature is enabled
+  (with HTTP traffic) PR [#973](https://github.com/metalbear-co/mirrord/pull/973).
 
 ## 3.20.0