+++ CHANGELOG.md
@@ -12,6 +12,11 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - Support for Go's `os.ReadDir` on Linux (by hooking the `getdents64` syscall). Part of
   [#120](https://github.com/metalbear-co/mirrord/issues/120).
 
+### Fixed
+
+- mirrord-agent: Handle HTTP upgrade requests when the stealer feature is enabled
+  (with HTTP traffic) PR [#973](https://github.com/metalbear-co/mirrord/pull/973).
+
 ## 3.20.0
 
 ### Added